api/app/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
psycopg2-binary==2.8.3
pytest-flask==0.14.0
requests==2.22.0
